monitor-manager: Evaluate adjusted value for overscan whitelist
When evaluating the output width and height, only the currently applied
values are considered. This poses a problem for when the overscanned
output reports a different size without the compensation borders.

Also, appearently, we cannot rely on the is_overscanned value to be
updated, since changing the XRandr property triggers multiple apply &
read cycles.

To fix the issue, a simpler approach was taken: simply check if either
the current monitor size or the overscan-adjusted monitor size match
any of the whitelisted sizes.

https://phabricator.endlessm.com/T13812