core: Add API to allow signal listeners to constrain geometry
This adds a new constraint to constraints.c which emits a signal
and, through the signal emission process, allows signal handlers
to expand the minimum size of the window. Attach to the
geometry-allocate signal and use meta_window_expand_allocated_geometry
to expand the minimum permissible width and height of the window.

Setting a smaller geometry is a no-op.